文章提交的功能终于全部写完了！

+++ src/main/java/com/codermy/myspringsecurityplus/fore/entity/ArticleTag.java
@@ -10,5 +10,5 @@ public class ArticleTag {
 
     private Long articleId;
 
-    private Long tagId;
+    private int tagId;
 }
